requirements.txt
streamlit
openai-whisper
librosa
numpy
scikit-learn
vaderSentiment
soundfile
torch